Cloud syncing

+++ src/services/supabase.js
@@ -1,45 +1,50 @@
 import { store } from "../store";
-import { supabase } from "../supabase"
+import { supabase } from "../supabase";
+import { compareHeroesProgressWithLS } from "./localStorage";
 
 export const completeSkill = async (hero, skill) => {
-  if (store.user?.id) {
-    const updatedHeroesProgress = store.heroesProgress;
+  const updatedHeroesProgress = store.heroesProgress;
 
-    if (!updatedHeroesProgress[hero]) {
-      updatedHeroesProgress[hero] = {};
-    }
-    updatedHeroesProgress[hero][skill] = true;
-
-    await supabase
-      .from("challenge")
-      .upsert({ heroes: updatedHeroesProgress, user_id: store.user.id }, { onConflict: "user_id" })
-      .match({ user_id: store.user.id });
+  if (!updatedHeroesProgress[hero]) {
+    updatedHeroesProgress[hero] = {};
   }
+  updatedHeroesProgress[hero][skill] = true;
+
+  await setHeroesProgressToSupabase(updatedHeroesProgress);
 };
 
 export const cancelSkill = async (hero, skill) => {
+  const updatedHeroesProgress = store.heroesProgress;
+
+  delete updatedHeroesProgress[hero][skill];
+
+  if (Object.values(updatedHeroesProgress[hero]).length === 0) {
+    delete updatedHeroesProgress[hero];
+  }
+
   if (store.user?.id) {
-    const updatedHeroesProgress = store.heroesProgress;
-
-    delete updatedHeroesProgress[hero][skill];
-
-    if (Object.values(updatedHeroesProgress[hero]).length === 0) {
-      delete updatedHeroesProgress[hero];
-    }
-
     if (Object.values(updatedHeroesProgress).length === 0) {
       await supabase
         .from("challenge")
         .delete()
         .match({ user_id: store.user.id });
     } else {
-      await supabase
-        .from("challenge")
-        .upsert({ heroes: updatedHeroesProgress, user_id: store.user.id }, { onConflict: "user_id" })
-        .match({ user_id: store.user.id });
+      await setHeroesProgressToSupabase(updatedHeroesProgress);
     }
   }
-}
+};
+
+export const setHeroesProgressToSupabase = async (heroesProgress) => {
+  if (store.user?.id) {
+    await supabase
+      .from("challenge")
+      .upsert(
+        { heroes: heroesProgress, user_id: store.user.id },
+        { onConflict: "user_id" }
+      )
+      .match({ user_id: store.user.id });
+  }
+};
 
 export const getHeroesProgressFromSupabase = async () => {
   if (store.user?.id) {
@@ -49,19 +54,20 @@ export const getHeroesProgressFromSupabase = async () => {
       .match({ user_id: store.user.id });
 
     if (Array.isArray(data)) {
-      store.heroesProgress = data[0]?.heroes ?? {};
+      const heroesProgress = data[0]?.heroes ?? {};
+      if (!compareHeroesProgressWithLS(heroesProgress)) {
+        store.cloud.showSyncModal = true;
+        store.cloud.data = heroesProgress;
+      }
     }
   }
 };
 
 export const resetChallengeProgress = async () => {
-  if (store.user?.id) {
-    await supabase
-      .from("challenge")
-      .delete()
-      .match({ user_id: store.user.id });
+  store.completedHeroes = {};
+  store.heroesProgress = {};
 
-    store.heroesProgress = {};
-    store.completedHeroes = {};
+  if (store.user?.id) {
+    await supabase.from("challenge").delete().match({ user_id: store.user.id });
   }
 };
